+++ utils/st_utils.py
@@ -1,19 +1,42 @@
+
+import os
+import re
+import time
+import base64
+from typing import List, Dict, Optional, Literal, Tuple
+
+from config.constants import (
+    I18N_DIR, 
+    SUPPORTED_LANGUAGES,
+    USER_AVATAR_SVG,
+    AI_AVATAR_SVG
+)
+from assets.styles.css.rag_chat_css import (
+    RAG_CHAT_USER_STYLE_ST_V37,
+    RAG_CHAT_USER_STYLE_ST_V39,
+    RAG_CHAT_ASSISTANT_STYLE_ST_V37,
+    RAG_CHAT_ASSISTANT_STYLE_ST_V39,
+)
+from assets.styles.css.classic_chat_css import (
+    USER_CHAT_STYLE_ST_V37,
+    USER_CHAT_STYLE_ST_V39,
+    ASSISTANT_CHAT_STYLE,
+)
+from assets.styles.css.export_themes import (
+    default_theme, 
+    glassmorphism_theme
+)
+from core.basic_config import I18nAuto
+from tools.toolkits import TO_TOOLS
+
 import streamlit as st
 import streamlit.components.v1 as components
-import os
-import time
 import whisper
-from typing import List, Dict, Optional, Literal
+import pyperclip
+import textwrap
 from streamlit_float import *
 from pkg_resources import parse_version
 
-from config.constants.i18n import I18N_DIR, SUPPORTED_LANGUAGES
-from core.basic_config import I18nAuto
-from utils.basic_utils import (
-    copy_to_clipboard,
-    export_chat_history_callback
-)
-from tools.toolkits import TO_TOOLS
 
 language = os.getenv("LANGUAGE", "简体中文")
 i18n = I18nAuto(
@@ -56,6 +79,26 @@ SCROLL_BUTTON_CONSTANTS = {
     }
 }
 
+# 定义样式常量
+STYLE_CONSTANTS = {
+    "USER_CHAT": {
+        "v37": USER_CHAT_STYLE_ST_V37,
+        "v39": USER_CHAT_STYLE_ST_V39,
+    },
+    "ASSISTANT_CHAT": {
+        "v37": ASSISTANT_CHAT_STYLE,
+        "v39": ASSISTANT_CHAT_STYLE,
+    },
+    "RAG_USER_CHAT": {
+        "v37": RAG_CHAT_USER_STYLE_ST_V37,
+        "v39": RAG_CHAT_USER_STYLE_ST_V39,
+    },
+    "RAG_ASSISTANT_CHAT": {
+        "v37": RAG_CHAT_ASSISTANT_STYLE_ST_V37,
+        "v39": RAG_CHAT_ASSISTANT_STYLE_ST_V39,
+    }
+}
+
 def get_scroll_button_js(
     scroll_type: Literal["BACK_TO_TOP", "BACK_TO_BOTTOM"],
     st_version: str
@@ -91,6 +134,413 @@ def get_scroll_button_collection(
         back_to_top=get_scroll_button_js(scroll_type="BACK_TO_TOP", st_version=st_version),
         back_to_bottom=get_scroll_button_js(scroll_type="BACK_TO_BOTTOM", st_version=st_version)
     )
+
+def copy_to_clipboard(content: str):
+    '''
+    将内容复制到剪贴板,并提供streamlit提醒
+    '''
+    pyperclip.copy(content)
+    st.toast(i18n("The content has been copied to the clipboard"), icon="✂️")
+
+def get_style(
+        style_type:Literal["USER_CHAT", "ASSISTANT_CHAT", "RAG_USER_CHAT", "RAG_ASSISTANT_CHAT"],
+        st_version:str
+    ) -> str:
+    """
+    根据样式类型和Streamlit版本获取相应的样式。
+    
+    :param style_type: 样式类型，如 "USER_CHAT", "ASSISTANT_CHAT" 等
+    :param st_version: Streamlit版本号, 形如 "1.37.0"
+    :return: 对应的样式字符串
+    """
+    version = parse_version(st_version)
+    style_dict = STYLE_CONSTANTS.get(style_type, {})
+    
+    if version < parse_version("1.38.0"):
+        return style_dict.get("v37", "")
+    else:
+        return style_dict.get("v39", "")
+
+
+def get_combined_style(
+        st_version:str, 
+        *style_types:Literal["USER_CHAT", "ASSISTANT_CHAT", "RAG_USER_CHAT", "RAG_ASSISTANT_CHAT"]
+    ) -> str:
+    """
+    获取多个样式类型的组合样式。
+    
+    :param st_version: Streamlit版本号, 形如 "1.37.0"
+    :param style_types: 要组合的样式类型列表
+    :return: 组合后的样式字符串
+    """
+    return "".join(get_style(style_type, st_version) for style_type in style_types)
+
+# @st.cache_data
+def write_chat_history(
+    chat_history: Optional[List[Dict[str, str]]] = None,
+    if_custom_css: bool = True
+) -> None:
+    # 将SVG编码为base64
+    user_avatar = f"data:image/svg+xml;base64,{base64.b64encode(USER_AVATAR_SVG.encode('utf-8')).decode('utf-8')}"
+
+    # 将SVG编码为base64
+    ai_avatar = f"data:image/svg+xml;base64,{base64.b64encode(AI_AVATAR_SVG.encode('utf-8')).decode('utf-8')}"
+
+    if chat_history:
+        for message in chat_history:
+            try:
+                if message["role"] == "system":
+                    continue
+            except:
+                pass
+            with st.chat_message(message["role"], avatar=user_avatar if message["role"] == "user" else ai_avatar):
+                st.html(f"<span class='chat-{message['role']}'></span>")
+                if isinstance(message["content"], str):
+                    st.markdown(message["content"])
+                elif isinstance(message["content"], List):
+                    for content in message["content"]:
+                        if content["type"] == "text":
+                            st.markdown(content["text"])
+                        elif content["type"] == "image_url":
+                            # 如果开头为data:image/jpeg;base64，则解码为BytesIO对象
+                            if content["image_url"]["url"].startswith("data:image/jpeg;base64"):
+                                image_data = base64.b64decode(content["image_url"]["url"].split(",")[1])
+                                st.image(image_data)
+                            else:
+                                st.image(content["image_url"])
+        
+        # 根据Streamlit版本选择样式
+        if if_custom_css:
+            chat_style = get_combined_style(st.__version__, "USER_CHAT", "ASSISTANT_CHAT")
+            st.html(chat_style)
+
+def wrap_long_text(text: str, max_length: int = 60) -> str:
+    """
+    将长文本按指定长度换行
+    """
+    wrapped_lines = textwrap.wrap(text, max_length)
+    return '<br>'.join(wrapped_lines)
+
+def generate_html_chat(
+        chat_history: List[Dict[str, str]], 
+        include_range: Optional[Tuple[int, int]] = None, 
+        exclude_indexes: Optional[List[int]] = None,
+        theme: Optional[str] = "default",
+        chat_name: Optional[str] = "Chat history",
+        model_name: Optional[str] = None,
+        code_theme: Optional[str] = "github-dark"
+    ) -> str:
+    """
+    生成HTML格式的聊天历史，支持Markdown渲染，并使用指定的主题
+
+    Args:
+        chat_history (List[Dict[str, str]]): 完整的聊天历史
+        include_range (Optional[Tuple[int, int]]): 要包含的消息索引范围，例如 (0, 10)
+        exclude_indexes (Optional[List[int]]): 要排除的消息索引列表，例如 [2, 5, 8]
+        theme (str, optional): 导出主题. Defaults to "default".
+        chat_name (str, optional): 聊天记录的名称. Defaults to "Chat history".
+        model_name (str, optional): 模型名称. Defaults to None.
+        code_theme (str, optional): 代码主题. Defaults to "github-dark".可选值见 https://highlightjs.org/examples
+
+    Returns:
+        str: 生成的HTML格式聊天历史
+    """
+    if theme == "default":
+        css_theme = default_theme
+    elif theme == "glassmorphism":
+        css_theme = glassmorphism_theme
+    else:
+        css_theme = default_theme  # 默认使用 default 主题
+
+    html_template = """
+    <!DOCTYPE html>
+    <html lang="zh-CN">
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <title>{chat_name}</title>
+        <script src="https://cdn.jsdelivr.net/npm/marked/marked.min.js"></script>
+        <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.7.0/styles/{code_theme}.min.css">
+        <script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.7.0/highlight.min.js"></script>
+        <style>
+            {css_theme}
+            pre {{
+                white-space: pre-wrap;
+                word-wrap: break-word;
+                background-color: #1A1B26;
+                border-radius: 5px;
+                padding: 1em;
+                margin: 1em 0;
+            }}
+            code {{
+                display: block;
+                overflow-x: auto;
+                padding: 0.5em;
+                background-color: #1A1B26;
+                font-family: 'Courier New', Courier, monospace;
+                color: #CBD2EA;
+            }}
+            code[class*="language-"] {{
+                background-color: #1A1B26;
+                color: #CBD2EA;
+            }}
+            .message.user pre,
+            .message.user code {{
+                text-align: left;
+            }}
+        </style>
+    </head>
+    <body>
+        <div class="card">
+            <div class="info-card">
+                <div class="info-left">
+                    <div class="project-name">RAGENT</div>
+                    <div class="project-url">
+                        <a href="https://github.com/Wannabeasmartguy/RAGENT" target="_blank">
+                            github.com/Wannabeasmartguy/RAGENT
+                        </a>
+                    </div>
+                </div>
+                <div class="info-right">
+                    <div class="info-right-content model-name">Model: {model_name}</div>
+                    <div class="info-right-content message-count">Messages: {message_count}</div>
+                    <div class="info-right-content chat-name">Chat: {chat_name}</div>
+                </div>
+            </div>
+            <div class="chat-container">
+                {chat_messages}
+            </div>
+        </div>
+        <script>
+            document.addEventListener('DOMContentLoaded', (event) => {{
+                marked.setOptions({{
+                    breaks: true,
+                    gfm: true,
+                    highlight: function(code, lang) {{
+                        const language = hljs.getLanguage(lang) ? lang : 'plaintext';
+                        return hljs.highlight(code, {{ language }}).value;
+                    }},
+                    langPrefix: 'hljs language-'
+                }});
+                
+                document.querySelectorAll('.markdown-content').forEach((element) => {{
+                    element.innerHTML = marked.parse(element.textContent);
+                }});
+                
+                hljs.highlightAll();
+
+                // 确保代码块内的换行符被保留
+                document.querySelectorAll('pre code').forEach((block) => {{
+                    block.innerHTML = block.innerHTML.replace(/\\n/g, '<br>');
+                }});
+            }});
+        </script>
+    </body>
+    </html>
+    """
+
+    model_name = model_name if model_name is not None else "Not specified"
+    wrapped_chat_name = wrap_long_text(chat_name)
+
+    if include_range is None:
+        include_range = (0, len(chat_history) - 1)
+    if exclude_indexes is None:
+        exclude_indexes = []
+    
+    # 计算消息数量，方法为：include_range范围长度 - exclude_indexes的元素个数
+    message_count = include_range[1] - include_range[0] + 1 - len(exclude_indexes)
+
+    chat_messages = []
+    for i in range(include_range[0], include_range[1] + 1):
+        if i in exclude_indexes:
+            continue
+
+        message = chat_history[i]
+        role = message['role']
+        content = message['content']
+        
+        message_html = f'<div class="message {role}">'
+        message_html += f'<div class="role">{role.capitalize()}</div>'
+        
+        if isinstance(content, str):
+            message_html += f'<div class="markdown-content">{content}</div>'
+        elif isinstance(content, list):
+            for item in content:
+                if item['type'] == 'text':
+                    message_html += f'<div class="markdown-content">{item["text"]}</div>'
+                elif item['type'] == 'image_url':
+                    image_url = item['image_url'].get('url', '')
+                    message_html += f'<img src="{image_url}" alt="Image">'
+        
+        message_html += '</div>'
+        chat_messages.append(message_html)
+    
+    # 使用 str.replace() 来插入 CSS 主题
+    html_content = html_template.replace("{css_theme}", css_theme)
+    # 然后使用 .format() 插入聊天消息
+    html_content = html_content.format(
+        chat_messages="\n".join(chat_messages),
+        chat_name=wrapped_chat_name,
+        model_name=model_name,
+        message_count=message_count,
+        code_theme=code_theme
+    )
+    
+    return html_content
+
+def generate_markdown_chat(
+    chat_history: List[Dict[str, str]], 
+    include_range: Optional[Tuple[int, int]] = None, 
+    exclude_indexes: Optional[List[int]] = None,
+    chat_name: Optional[str] = "Chat history"
+) -> str:
+    """
+    生成Markdown格式的聊天历史
+
+    Args:
+        chat_history (List[Dict[str, str]]): 完整的聊天历史
+        include_range (Optional[Tuple[int, int]]): 要包含的消息索引范围，例如 (0, 10)
+        exclude_indexes (Optional[List[int]]): 要排除的消息索引列表，例如 [2, 5, 8]
+
+    Returns:
+        str: 生成的Markdown格式聊天历史
+    """
+    formatted_history = []
+    image_references = []
+    image_counter = 0
+
+    # 如果没有指定范围，则处理所有消息
+    if include_range is None:
+        include_range = (0, len(chat_history) - 1)
+    
+    # 如果没有指定排除索引，初始化为空列表
+    if exclude_indexes is None:
+        exclude_indexes = []
+
+    formatted_history.append(f"# {chat_name}\n\n")
+
+    for i in range(include_range[0], include_range[1] + 1):
+        if i in exclude_indexes:
+            continue
+
+        message = chat_history[i]
+        role = message['role'].title()
+        content = message['content']
+        
+        formatted_history.append(f"## {role}\n\n")
+        
+        if isinstance(content, str):
+            formatted_history.append(f"{content}\n\n")
+        elif isinstance(content, list):
+            for item in content:
+                if item['type'] == 'text':
+                    formatted_history.append(f"{item['text']}\n\n")
+                elif item['type'] == 'image_url':
+                    image_url = item['image_url'].get('url', '')
+                    image_counter += 1
+                    reference_id = f"image{image_counter}"
+                    
+                    if image_url.startswith('data:image/jpeg;base64,'):
+                        formatted_history.append(f"![image][{reference_id}]\n\n")
+                        image_references.append(f"[{reference_id}]: {image_url}\n")
+                    else:
+                        formatted_history.append(f"![Image]({image_url})\n\n")
+    
+    # 添加图片引用到文档末尾
+    if image_references:
+        formatted_history.append("\n\n<!-- Image References -->\n")
+        formatted_history.extend(image_references)
+    
+    chat_history_text = "".join(formatted_history)
+
+    return chat_history_text
+
+def export_chat_history_callback(
+        chat_history: List[Dict[str, str]], 
+        include_range: Optional[Tuple[int, int]] = None,
+        exclude_indexes: Optional[List[int]] = None,
+        is_rag: bool = False,
+        export_type: Optional[str] = "html",
+        theme: Optional[str] = "default",
+        chat_name: Optional[str] = "Chat history",
+        model_name: Optional[str] = None
+    ):
+    """
+    导出聊天历史记录
+    
+    Args:
+        chat_history (List[Dict[str, str]]): 聊天历史记录
+        include_range (Optional[Tuple[int, int]]): 要包含的消息索引范围，例如 (0, 10)
+        exclude_indexes (Optional[List[int]]): 要排除的消息索引列表，例如 [2, 5, 8]
+        is_rag (bool, optional): 是否是RAG聊天记录. Defaults to False.
+        export_type (str, optional): 导出类型，支持 "markdown" 和 "html". Defaults to "html".
+        theme (str, optional): 导出主题. Defaults to "default".仅当export_type为html时有效
+        chat_name (str, optional): 聊天记录的名称. Defaults to "Chat history".
+        model_name (str, optional): 模型名称. Defaults to None.仅当export_type为html时有效
+    """
+    # 清理文件名，移除非法字符
+    chat_name = re.sub(r'[\\/*?:"<>|]', "", chat_name).strip()
+    if not chat_name:
+        chat_name = "Chat history"
+
+    export_folder = "chat histories export"
+    os.makedirs(export_folder, exist_ok=True)
+
+    if export_type == "markdown":
+        markdown_content = generate_markdown_chat(
+            chat_history=chat_history,
+            include_range=include_range,
+            exclude_indexes=exclude_indexes,
+            chat_name=chat_name
+        )
+
+        filename = f"{'RAG ' if is_rag else ''}Chat history - {chat_name}.md"
+        i = 1
+        while os.path.exists(os.path.join(export_folder, filename)):
+            filename = f"{'RAG ' if is_rag else ''}Chat history - {chat_name} ({i}).md"
+            i += 1
+
+        full_path = os.path.join(export_folder, filename)
+        with open(full_path, "w", encoding="utf-8") as f:
+            f.write(markdown_content)
+        st.toast(body=i18n(f"Chat history exported to: {full_path}"), icon="🎉")
+    
+    elif export_type == "html":
+        html_content = generate_html_chat(
+            chat_history=chat_history,
+            include_range=include_range,
+            exclude_indexes=exclude_indexes,
+            theme=theme,
+            chat_name=chat_name,
+            model_name=model_name
+        )
+
+        filename = f"{'RAG ' if is_rag else ''}Chat history - {chat_name}.html"
+        i = 1
+        while os.path.exists(os.path.join(export_folder, filename)):
+            filename = f"{'RAG ' if is_rag else ''}Chat history - {chat_name} ({i}).html"
+            i += 1
+
+        full_path = os.path.join(export_folder, filename)
+        with open(full_path, "w", encoding="utf-8") as f:
+            f.write(html_content)
+        st.toast(body=i18n(f"Chat history exported to: {full_path}"), icon="🎉")
+
+    # elif export_type == "jpg":
+    #     image_stream = html_to_jpg(html_content)
+    #     export_folder = "chat histories export"
+    #     filename = "RAG Chat history.jpg" if is_rag else "Chat history.jpg"
+    #     i = 1
+    #     while os.path.exists(os.path.join(export_folder, filename)):
+    #         filename = f"{'RAG ' if is_rag else ''}Chat history({i}).jpg"
+    #         i += 1
+        
+    #     os.makedirs(export_folder, exist_ok=True)
+    #     with open(os.path.join(export_folder, filename), "wb") as f:
+    #         image_stream.save(f, format="JPEG")
+    #     st.toast(body=i18n(f"Chat history exported to: " + os.path.join(export_folder, filename)), icon="🎉")
+    else:
+        st.error(i18n("Unsupported export type"))
 
 def back_to_top(script_container = st.empty(), buttom_container = st.empty()):
     """
